Move navigation menu logic to trait
